limpia de código, refactorización y paso a inglés

+++ src/main/java/com/example/demo/model/service/ProductsService.java
@@ -30,7 +30,7 @@ public class ProductsService {
 		return productoDAO.findAll();
 	}
 
-	public Product searchProduct(int id) {
+	public Product searchProduct(Long id) {
 		Optional<Product> opP;
 		opP = productoDAO.findById(id);
 		if (opP.isPresent()) {
